+++ nurbs/src/mesh.h
@@ -1,13 +1,13 @@
 #pragma once
 #include <vector>
 #include <string>
-#include <gismo/gismo.h>
+#include <gismo.h>
 
 using namespace gismo;
 
 struct mesh
 {
-    std::vector<gsVector3d<double>> vertices;
+    std::vector<gsVector3d<float>> vertices;
 
     void to_stl(const std::string &filename) const;
