+++ nurbs/src/nurbs_refine.h
@@ -3,6 +3,6 @@
 #include "nurbs_include.h"
 
 template <size_t dimension>
-double nurbs<dimension>::refine(size_t dim, size_t index, double x, size_t deg) const
+float nurbs<dimension>::refine(size_t dim, size_t index, float x, size_t deg) const
 {
 }
