+++ nurbs/src/transformations.h
@@ -5,6 +5,6 @@
 template <>
 int nurbs<3>::bend(
     size_t param_dim,
-    double param_value,
-    double angle,
-    double radius);
+    float param_value,
+    float angle,
+    float radius);
